+++ eth_defi/vault/risk.py
@@ -70,6 +70,7 @@ VAULT_PROTOCOL_RISK_MATRIX = {
     "Peapods": VaultTechnicalRisk.dangerous,
     "Ostium": VaultTechnicalRisk.high,
     "gTrade": VaultTechnicalRisk.high,
+    "Domination Finance": VaultTechnicalRisk.severe,
     # No audits
     "Plutus": VaultTechnicalRisk.severe,
     "Harvest Finance": VaultTechnicalRisk.low,